Minor changes to move StudyDeck
